requirements.txt
onnxruntime
nemo_toolkit[asr]
torchaudio
fastapi
uvicorn
torch
